Add support for multiple terminal runs
Inside any terminal, go to visual mode and highlight some text (probably
a command you just ran), then hit <leader>rr

Now in any terminal in normal mode <leader>rr will repeat that command
in the same terminal.

This works for
  - <leader>rr
  - <leader>rd
  - <leader>rt

Which is useful eg for having a command for running tests and another
for debugging a test.