improve test + repl experience
h#EditREPL edits a repl file and opens a repl in the top left.  Writes
to this file automatically get sent to the repl terminal.

switch vim-test config to use a custom neoterm strategy that always
sends to the bottom left, opening a neoterm there if needed.

Update workspace to use neoterm.  As part of this, do not automatically
startinsert in terminals (but still do so on WinEnter).  In practice
this doesn't change much since Tnew leaves the cursor in the starting
window.  It fixes lots of annoying edge cases though, because
startinsert is effectively async.

Update workspace sourcing to automatically create a .vimrc if it's
missing, by copying a blueprint.  The blueprint will have some useful
project-specific setting commented out to show how to have
project-specific env vars for testing, and how to sensibly invoke
h#EditREPL.