Change main() to be opt-in
This is only partially done (for keymaps and plugin configuration), but
it allows tests to run against only a subset of main() (e.g. configuring
a specific plugin).

Note that whether a plugin is "loaded" depends on the plugin -- they're
all installed into `/start` and not `/opt`, however some plugins will be
largely inert until configured.