Autocomplete spike for vim APIs
Sumneko doesn't understand where to find types for APIs like
`vim.keymap`. However it upvalues types in a fairly simplistic way, so a
simple pattern of overwritting vim.keymap within an if false block gets
autocomplete w/docs working pretty well.

Autocomplete + docs will also work for importing functions whose docs
are copied from C. This isn't quite as trivial as it isn't a simple
require, but it's something that could be generated for any particular
version of neovim.

The general idea needs to be fleshed out more but it would make a nice
plugin.