feat: Improve terminal usage
Instead of having a terminal on the left, have two kinds of terminals:

* The floating terminal from ctrl-t
* Terminals in a separate tab for long-running processes (e.g. servers)

To make this more convenient, configure kitty + neovim so that ⌘-n goes
to tab n and ⌘-t creates a new (neovim) tab.  This prevents me from
using kitty tabs, which is fine since I never want to do that.