fix: disable proc-macro-disabled diagnostic
rust-analyzer was showing me a diagnostic for disabled proc macros that
would highlight the entire function the macro applied to.  That was very
distracting.

I have no idea why the proc macro was even disabled, but this is the
shortest path to not dealing with the issue.

It should work to specify

```lua
["rust-analyzer"] = {
  procMacro = { enable = true }
}
```

But that didn't seem to work for me.