feat: improve noice config
Add a debugging flag (and keymap toggle) that causes most messages to
get output & available via `<leader>sna`.  This makes it much easier to
figure out how to configure routes for messages.

As part of this, also update noice routes to greatly simplify the way in
which I silence the annoying toast messages.