fix(formatting: python
In the first place, use `ruff_format` and not `ruff`.

In the second place, do not allow mason to set the PATH ever.  Do it
manually in `zshrc`.  Mason has an option for appending the path but it
doesn't work.  I need mason to be last so that projects that are very
picky about versions (of e.g. ruff) are allowed to set the ruff version.